+++ src/backend/mod.rs
@@ -20,9 +20,9 @@ use core::fmt::Formatter;
 use std::fmt::Display;
 
 use bpf::{
-    ARG_NUMBER_MAX, AUDIT_ARCH_AARCH64, AUDIT_ARCH_X86_64, BPF_MAX_LEN, SECCOMP_RET_ALLOW,
-    SECCOMP_RET_ERRNO, SECCOMP_RET_KILL_PROCESS, SECCOMP_RET_KILL_THREAD, SECCOMP_RET_LOG,
-    SECCOMP_RET_MASK, SECCOMP_RET_TRACE, SECCOMP_RET_TRAP,
+    ARG_NUMBER_MAX, AUDIT_ARCH_AARCH64, AUDIT_ARCH_RISCV64, AUDIT_ARCH_X86_64, BPF_MAX_LEN,
+    SECCOMP_RET_ALLOW, SECCOMP_RET_ERRNO, SECCOMP_RET_KILL_PROCESS, SECCOMP_RET_KILL_THREAD,
+    SECCOMP_RET_LOG, SECCOMP_RET_MASK, SECCOMP_RET_TRACE, SECCOMP_RET_TRAP,
 };
 
 pub use bpf::{sock_filter, BpfProgram, BpfProgramRef};
@@ -81,6 +81,8 @@ pub enum TargetArch {
     x86_64,
     /// aarch64 arch
     aarch64,
+    /// riscv64 arch
+    riscv64,
 }
 
 impl TargetArch {
@@ -89,6 +91,7 @@ impl TargetArch {
         match self {
             TargetArch::x86_64 => AUDIT_ARCH_X86_64,
             TargetArch::aarch64 => AUDIT_ARCH_AARCH64,
+            TargetArch::riscv64 => AUDIT_ARCH_RISCV64,
         }
     }
 }
@@ -99,6 +102,7 @@ impl TryFrom<&str> for TargetArch {
         match input.to_lowercase().as_str() {
             "x86_64" => Ok(TargetArch::x86_64),
             "aarch64" => Ok(TargetArch::aarch64),
+            "riscv64" => Ok(TargetArch::riscv64),
             _ => Err(Error::InvalidTargetArch(input.to_string())),
         }
     }
@@ -202,6 +206,15 @@ mod tests {
         assert_eq!(
             TargetArch::try_from("aARch64").unwrap(),
             TargetArch::aarch64
+        );
+
+        assert_eq!(
+            TargetArch::try_from("riscv64").unwrap(),
+            TargetArch::riscv64
+        );
+        assert_eq!(
+            TargetArch::try_from("RiScV64").unwrap(),
+            TargetArch::riscv64
         );
     }
 }